move data loader code, split out logic where it makes sense
