+++ pkg/storage/loaders/data_loader_middleware.go
@@ -7,11 +7,7 @@ import (
 	"net/http"
 
 	"github.com/graph-gophers/dataloader"
-	"github.com/samber/lo"
-	"go.uber.org/zap"
 
-	"github.com/cmsgov/mint-app/pkg/appcontext"
-	"github.com/cmsgov/mint-app/pkg/models"
 	"github.com/cmsgov/mint-app/pkg/storage"
 )
 
@@ -130,93 +126,10 @@ type DataReader struct {
 	Store *storage.Store
 }
 
-// GetPlanBasicsByModelPlanID uses a DataLoader to aggreggate a SQL call and return all plan basics in one query
-func (loaders *Loaders) GetPlanBasicsByModelPlanID(ctx context.Context, keys dataloader.Keys) []*dataloader.Result {
-	dr := loaders.DataReader
-
-	modelPlanIDs := make([]string, len(keys))
-	for ix, key := range keys {
-		modelPlanIDs[ix] = key.String()
-	}
-	logger := appcontext.ZLogger(ctx)
-	basics, _ := dr.Store.PlanBasicsGetByModelPlanIDLOADER(logger, modelPlanIDs)
-
-	// RETURN IN THE SAME ORDER REQUESTED
-	output := make([]*dataloader.Result, len(keys))
-	for index, key := range keys {
-		basic, ok := lo.Find(basics, func(basic *models.PlanBasics) bool { //Get the  plan basic that matches what we asked for
-
-			return basic.ModelPlanID.String() == key.String()
-		})
-
-		if ok {
-			output[index] = &dataloader.Result{Data: basic, Error: nil}
-		} else {
-			err := fmt.Errorf("plan basic not found for model plan %s", key.String())
-			output[index] = &dataloader.Result{Data: nil, Error: err}
-		}
-	}
-	return output
-
-}
-
 func stringArrayFromKeys(keys dataloader.Keys) []string {
 	stringArr := make([]string, len(keys))
 	for ix, key := range keys {
 		stringArr[ix] = key.String()
 	}
 	return stringArr
-}
-
-// GetOperationalNeedsByModelPlanID uses a data loader to aggregate SQL calls and return data
-func (loaders *Loaders) GetOperationalNeedsByModelPlanID(ctx context.Context, keys dataloader.Keys) []*dataloader.Result {
-
-	modelPlanIDs := stringArrayFromKeys(keys)
-	logger := appcontext.ZLogger(ctx)
-	dr := loaders.DataReader
-
-	opNeeds, _ := dr.Store.OperationalNeedCollectionGetByModelPlanIDLOADER(logger, modelPlanIDs)
-
-	// RETURN IN THE SAME ORDER REQUESTED
-	output := make([]*dataloader.Result, len(keys))
-	for index, key := range keys {
-		needs := lo.Filter(opNeeds, func(opNeed *models.OperationalNeed, index int) bool {
-			return opNeed.ModelPlanID.String() == key.String()
-		})
-		output[index] = &dataloader.Result{Data: needs, Error: nil}
-
-	}
-	return output
-
-}
-
-// GetOperationalSolutionAndPossibleCollectionByOperationalNeedID uses a data loader to return operational solutions by operational need id
-func (loaders *Loaders) GetOperationalSolutionAndPossibleCollectionByOperationalNeedID(ctx context.Context, keys dataloader.Keys) []*dataloader.Result {
-	logger := appcontext.ZLogger(ctx)
-	arrayCK, err := CompoundKeyArray(keys)
-	if err != nil {
-		logger.Error("issue converting keys for data loader in Operational Solutions", zap.Error(*err))
-	}
-	marshaledParams, err := arrayCK.ToJSONArray()
-	if err != nil {
-		logger.Error("issue converting keys to JSON for data loader in Operational Solutions", zap.Error(*err))
-	}
-
-	dr := loaders.DataReader
-
-	sols, loadErr := dr.Store.OperationalSolutionAndPossibleCollectionGetByOperationalNeedIDLOADER(logger, marshaledParams)
-
-	output := make([]*dataloader.Result, len(keys))
-	for index, key := range keys {
-		ck := key.Raw().(CompoundKey)
-		opNeedID := fmt.Sprint(ck.Args["operational_need_id"])
-		needs := lo.Filter(sols, func(opSol *models.OperationalSolution, index int) bool {
-
-			return opSol.OperationalNeedID.String() == opNeedID
-		})
-		output[index] = &dataloader.Result{Data: needs, Error: loadErr}
-
-	}
-	return output
-
 }
